requirements.txt
aiosqlite>=0.17.0
asyncpg>=0.29.0
aiofiles>=23.2.1
PyPDF2>=3.0.0
python-docx>=0.8.11
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.25.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'
python-dateutil>=2.8.0
matplotlib>=3.7.1
bcrypt>=4.0.0
pandas>=2.0.0
cohere
python-jose[cryptography]
openai
streamlit-autorefresh